auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Juanpfrancos <juanpanime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
pydantic = {extras = ["email"], version = "^2.10.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
